+++ Calc.Compiler/IOperator.cs
@@ -8,6 +8,6 @@
 		/// <param name="operand1">Operando 1.</param>
 		/// <param name="operand2">Operando 2.</param>
 		/// <returns></returns>
-		float Calculate(float operand1, float operand2);
+		double Calculate(double operand1, double operand2);
 	}
 }
